Support "unambiguous" sourceType option
Summary: The Babel parser supports an option for `sourceType` that the WASM Hermes parser does not - unambiguous, which means a `sourceType` of `"module"` or `"script"` will be detected based on the presence of ES6 imports/exports. The Hermes parser already has this functionality when no `sourceType` is provided. Let's simply allow for passing in an `{sourceType: "unambiguous"}` to the Hermes parser for consistency with Babel's API.

Reviewed By: pieterv

Differential Revision: D26649384

fbshipit-source-id: 55a852a306cfbaa3f8dcd852bf907789ab58fd46