Use upper_bound for lastDefBeforeFirstUse
Summary:
In order to find the last def that occurs at or before the first use,
we can use a `std::set` and `std::set::upper_bound`. This is clearer
and fixes a bug where we currently do not consider defs that occur at
the first use (instead only considering defs before it).

Reviewed By: dulinriley

Differential Revision: D26697839

fbshipit-source-id: 1f1cd5c6a96cf4f3bd799e53e37520db5df182de