Fix heap timeline serialization bugs
Summary:
Fixes a number of related correctness issues in heap timeline serialization:
1. We were accessing `AllocationLocationTracker::fragments_` without locking the associated mutex - a potential problem under Hades.
2. We were not flushing the current fragment before serializing, potentially missing some allocations at the tail end of the timeline.
3. We were appending an invalid fragment at the end of the `samples` array. When Chrome loads this data from a `.heaptimeline` file (and not directly from the debugger), the timeline is truncated at 10s and allocations after that point are often missing entirely (e.g. from Summary view).

To fix this holistically, we move the loop over `fragments_` into `AllocationLocationTracker` where we can (1) lock the mutex, (2) flush the current fragment if necessary, and (3) limit serialization to fragments that have been flushed.

Note that only (1) and (2) are related to serializing while allocation tracking is enabled - (3) is *always* the case even if tracking is disabled first.

Reviewed By: dulinriley

Differential Revision: D26676168

fbshipit-source-id: e51b7d479579e89f45107bdac5dfc5a32e61d466