Make card table atomic in Hades
Summary:
Hades relies on being able to concurrently dirty cards from write
barriers and the background thread. To make sure this is done safely,
we can make the cards atomic.

An ideal implementation would be to use a C++20 `std::atomic_ref` in
`dirtyCardForAddress` and keep the underlying data as non-atomic.
That would also be safe because `dirtyCardForAddress` is the only
concurrent access (by the write barriers and the marking thread).
Then we would be able to use memchr and friends to make
manipulating cards faster.

Reviewed By: dulinriley

Differential Revision: D26712399

fbshipit-source-id: bfed8f292760f9f055edef5e7627ffcc8f0c8fb3