Remove alignment requirements for heap segment metadata
Summary:
Remove the need for the card table and mark bit array to be page
aligned. This lets us reduce the amount of unnecessary padding when
the heap segment size is small.

This diff is focused on adding padding to align the guard page and
adjusting asserts so that they reflect the new layout. The next diff
will remove the alignment requirements from the card table itself.

For instance, consider the case where the heap segment size is set to
256KB. The card table is then only 512 bytes. However, the card table
and the COT are each aligned to a page, which is 4KB on Android. So
we end up with 8KB worth of space for 1KB worth of data. When we add
in the guard page and the mark bit array (one page each), we end up
with a total of 16KB overhead per segment. Removing alignment
requirements for the card table reclaims 4KB, which is significant.

Reviewed By: dulinriley

Differential Revision: D26716505

fbshipit-source-id: f112d4b996a3d1f96015037b08647ed11389bd7b