Make heap segment size configurable
Summary:
The segment size is a very useful lever that we should probably make
use of more. For instance, it can be used to:
1. Reduce YG pause times, since YG is at most 1 segment in size. This
will also help improve pause times in incremental Hades.
2. Work with small heaps (<16MB). Since Hades in particular does not
work with sub-segment heaps, having a smaller segment size is the
best way to work with small heaps.

Reviewed By: dulinriley

Differential Revision: D26713861

fbshipit-source-id: c0595018c8a848eec9bf0cfc3d66084f9b18061e