Fix flakiness in interpreter stack frame size test
Summary:
This test is currently flaky because the outer stack pointer value is
completely dependent on how the compiler decides to lay out the
stack. Instead, force the stack size measurement into a separate,
non-inline function so that the compiler is limited in how it can
reorganise the stack.

This helps ensure that the `dummy` variable doesn't get moved before
other stack variables, resulting in us incorrectly measuring a larger
stack.

Reviewed By: tmikov

Differential Revision: D26792559

fbshipit-source-id: 21f7711de47398803060f2e93ecf42356853e195