Fix synth trace builds with GCC
Summary:
We currently cannot build synth trace tests and libraries with GCC.

`SynthTraceParser.cpp` relies on LLVHSupport, which is built without
RTTI. This means that `SynthTraceParser.cpp` itself has to be built
without RTTI. However, it does use exceptions so we have to create a
new special case for it where exceptions are on but RTTI is off.

The synth trace tests are somewhat more complicated. They rely on
LLVH, RTTI, exceptions, and the hermes API simultaneously. Getting
them to build reliably will be much more challenging, and may require
refactoring them to depend on folly JSON instead of the JSON parser
in Hermes, or untangling all of these features somehow. Either way,
for now, I propose removing them from the OSS tests, and only testing
them in buck, like we do for the heap snapshot tests.

Reviewed By: dulinriley

Differential Revision: D26809542

fbshipit-source-id: 35b00f6e4385ef816e381c73dd8ee28646782907