Flow type definition files for exports of hermes-parser and hermes-eslint
Summary:
Let's add Flow type definition files for the exports of `hermes-parser` and `hermes-eslint`, so that users can use them in a more type safe manner. These definitions will live in `index.js.flow`, and contain types for the exported `parse` and `parseForESLint` functions.

At the moment this is mostly useful for specifying the allowed arguments. We do not yet have type definitions for the scope manager or the AST, so we leave these as generic Object types (aliased to a more helpful name - Program or ScopeManager).

Reviewed By: pieterv

Differential Revision: D26990834

fbshipit-source-id: ec4150245c3b15d19593b628b23ef9dc14de91f7