Prevent changes in cell value after null check
Summary:
We rely on only reading heap values once in Hades acceptors, but not
on having the most up-to-date value. For instance, after reading a
GCHermesValue and checking whether it is a pointer, we rely on that
value continuing to be a valid pointer.  For GCPointers, we rely on
the value not becoming null after a null-check.

To ensure that we only read from these memory locations once, cast
them to volatile pointers before the read. Since reads from volatile
variables are considered observable behaviour, the compiler cannot
optimise away the local variable in favour of reading multiple times.

Reviewed By: dulinriley

Differential Revision: D26417378

fbshipit-source-id: a33f186fb1dd7c2f445913d5efcebfb6b66dbf4d