Make opcode zero be unreachable and fatal the process
Summary:
Currently the opcode whose value is zero is `NewObjectWithBuffer`.
When reading random bytes out of memory, 0 tends to be more common than
other values.

This means if something weird happens to the IP and it gets pointed somewhere
random in memory, it has a high chance of reading what it thinks is a valid instruction.

Change this so 0 is a special instruction that fatals the process. Repurpose the `Unreachable`
instruction for this, as nothing else was using it.

The bytecode version is bumped because all other instructions are now one higher.

Reviewed By: tmikov

Differential Revision: D26994278

fbshipit-source-id: 39f80a3dd2c92dbb26dcd9a962bb63f27054e9bd