Implement a sampling memory profiler
Summary:
The Chrome sampling memory profiler works as follows:
* At a random interval of allocated bytes, take a sample
* A sample includes the size of the allocation and the stack that allocated that sample
* If the object is collected, so is the sample. Thus it only tracks objects that live until the end of recording
* Once recording is finished, a list of samples and their stacks are returned to the inspector

Chrome itself uses a Poisson distribution whose mean is the sampling interval, to avoid
sampling in a predictable rhythm. Hermes copies that same idea.

The output format is a JSON file with two top-level keys, a "head" that describes a call
stack tree, and a "samples" array that defines what each sample is, and which stack node
it was allocated at.

These semantics match up very closely to the AllocationLocationTracker, so copy the same
API: `newAlloc`, `updateSize`, `freeAlloc`, and merge them all into GC APIs. It's possible for
both to be enabled at once in this implementation, but Chrome doesn't support that, so it's
not an important requirement. It also re-uses the StackTraceTree to ensure the stack
reporting is always correct.

Since the output format shares almost nothing with other snapshot mechanisms, define
a new class `ChromeSamplingMemoryProfile` to handle serialization to JSON. It's much
simpler than the non-sampling memory profiler.

Reviewed By: neildhar

Differential Revision: D25789146

fbshipit-source-id: 856cf889ef24fc6622fc26a8c3c8f1760007dff0