Add the sampling heap profiler as an action to synth benchmarks
Summary:
To further test the sampling heap profiler, and allow runners of synth
benchmarks to get some flame graphs of memory allocations,
add a new action to synth benchmarks to take a sampling memory profile.

This can be done with `-action-at-marker=sample` which means to stop sampling
and get the results once the marker is hit.

Reviewed By: neildhar

Differential Revision: D27086260

fbshipit-source-id: 92a384b1400f8b003bfed79a2f767757229dc6b7