Refactor JSObject named descriptor API.
Summary:
In practice, the JSObject named descriptor API is used in two cases:

1. The name of the property is known when compiling the VM.
   In this case, predefine the string and use the new descriptor
   functions which operate on Predefined: `getNamedDescriptorPredefined`.
2. The name of the property is in the bytecode string table.
   In this case, the IdentifierTable will keep the `SymbolID` alive,
   so JSObject doesn't have to do extra work to keep it alive.
   Introduce a `getNamedDescriptorUnsafe` which operates on `SymbolID`s
   which are known to be kept alive for the duration of the function call.

To handle Proxy and HostObject cases, separate the functions
which take the descriptor into standard and `Internal` variants,
to be used when the target object can be a Proxy or HostObject
and definitely can't, respectively.

Users must only call the `Internal` variant when it is known that the
object cannot be a Proxy or HostObject.

Reviewed By: tmikov

Differential Revision: D24434878

fbshipit-source-id: 90a5d408159b83fa6984ea5cfd927c0ae634c9b0