Make Hades YG 2MB
Summary:
In order to reduce pause times, make the YG allocation area in Hades
2MB.

There are a few additional considerations handled in this diff:
1. During YG promotion, keep the YG segment size as 4MB. This avoids
wasting space in the OG by having partially full segments. There is
also no pause time benefit to having 2MB YG in during promotion.
2. If we're making an allocation >2MB, temporarily increase the size
of the YG. We can then adjust it back at the end of a YG collection.
The alternative is to directly allocate that object in the OG, but I
chose not to do that in the interest of minimising direct to OG
allocations, particularly since this is such a rare case.
3. In future, we may want to adjust this size dynamically, based on
the measured YG times. Then we could dynamically trade off the
latency and throughput.

Reviewed By: dulinriley

Differential Revision: D27245952

fbshipit-source-id: 961daae111501b4c48ddb98a6888c2c532679f1b