Store JSArray indexedStorage pointer in a direct property slot
Summary:
Instead of storing the indexedStorage in a GCPointer, store it in a
direct property slot in the underlying JSObject. This simplifies the
layout of JSArray, by making its size an integer multiple of 8 bytes.
That in turn means that the compiler does not insert padding in
JSArray, avoiding false overlap with direct properties.

With HV32, having this false overlap would prevent us from fitting
`length` in a direct property slot.

Reviewed By: dulinriley

Differential Revision: D27355592

fbshipit-source-id: c063735a9ca975bb0109af415572915060f3c708