Fix race condition in assert with WeakRefSlot::value and others
Summary:
There were some methods in `WeakRefSlot` that checked the `state`
while it could be modified by the marking thread in the background.
Since these are just asserts they're not actually production issues requiring a lock or atomics,
but we should fix them to prevent race conditions in dev/dbg builds.

Reviewed By: neildhar

Differential Revision: D27378002

fbshipit-source-id: 975b2b18fd807f60a57e07c5598a2aaf52322726