Allocate prop storage for additional slots if necessary
Summary:
`DecoratedObject` and `NativeFunction` support adding a dynamic
number of "additional slots", which are slots reserved in the base
`JSObject` and hidden class for internal use. In theory, if the
number of additional slots is large enough, or if overlap slots use
up all of the direct slots, these slots could end up in indirect prop
storage.

However, we currently do not allocate prop storage for these
additional slots, which implicitly assumes that they will always fit
in direct storage. With HV32, the size of direct props will shrink,
resulting in more direct slots being lost to overlap. This will
create real cases where additional slots need to use indirect prop
storage.

To fix this, allocate the appropriate amount of prop storage in the
respective `create` functions. This is already done in some overloads
of `JSObject::create`.

Reviewed By: dulinriley

Differential Revision: D27340082

fbshipit-source-id: 5c3059e3de8c3f010499356c8c34308f22e8fff1