Avoid iterator invalidation in eraseSegmentFreelists
Summary:
In the current implementation of this loop, we increment the iterator
at the end. However, if the call to `reset(idx)` results in the
current node of the SparseBitVector being freed, then the iterator
will get invalidated. Instead, we should increment the iterator at
the start of the body, so that invalidating the iterator associated
with `idx` does not cause a crash.

This works because SparseBitVector is internally implemented as a
std::list of nodes, and each node stores 128 bits. So invalidating
the iterator to a single node still keeps the subsequent iterators
valid.

This bug can only show up at node boundaries, because we first set
`idx-1` before resetting `idx`. So for instance, if `idx` were 128
exactly, then the call to `set(idx-1)` would happen in the previous
node, but the call to `reset(idx)` would happen in the current node.
If there is nothing else in the current node, then it gets freed, and
attempting to advance the iterator causes a UAF.

Reviewed By: avp

Differential Revision: D27482986

fbshipit-source-id: 2881f288ff12ec77312b383f124528477056b131