Add a DiagnosticStat type to generalise tracking element sizes
Summary:
In order to make the heap size diagnostic less error prone, introduce
a new generic DiagnosticStat type that we can use to track stats and
print results. This will also make it easier to port the code into
GCBase so that it can be shared across GCs.

In this diff, switch StringPrimitive handling into this new acceptor.

Reviewed By: dulinriley

Differential Revision: D27458506

fbshipit-source-id: b4c18f0ffe6700ffbcea2fd3b288f23fac73644d