Add documentation for Hades
Summary:
Hades is now a fully baked garbage collector that's going to become the
default soon (anticipated with Hermes v0.8).
Include some documentation about how it works.

A good chunk of the design is the same as GenGC, so this documentation links
there and tries to address the differences, namely the allocation pattern
in the Old Gen, and the concurrent nature of an Old Gen GC.

It explains many of the design choices, most notably the Snapshot at the Beginning
(SATB) write barrier that is the key to the implementation.

Reviewed By: neildhar

Differential Revision: D27632157

fbshipit-source-id: 7924cd752068582e6f75408c251e12044a4737d4