Length comparison incorrect for TypedArray.from (#493)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/493

Fix #485.

The `length` can be any integer from 0 to 2 ^ 53 - 1, so it can't be cast to a uint32_t
safely. Doing so hits an assert in `getNumberAs<uint32_t>`.
Change the type signature to `uint64_t`, since this is only ever called with numbers anyway.

After this change, the added unit test throws a TypeError instead of hitting an assert.

Reviewed By: avp

Differential Revision: D27683567

fbshipit-source-id: f194e73e3031f2ed3bd141132ae32950540fd6b2