Dynamically resize Hades YG
Summary:
Now that we know that a smaller YG can significantly improve pause
times on lower end devices, try dynamically resizing the YG so that
we can determine how to meet pause time goals at runtime. This lets
us go even lower on low end devices (I've currently set a floor of
1MB) without losing throughput on higher end devices due to increased
collection frequency.

Note that this is premised on the assumption that the tail end of
slow collections we see happen on the same devices, and that there
isn't huge variation in YG collection device on a single device.

Start at a 50% ratio, since this is half the max and 2x the minimum,
which means that we can get to either extreme within 6-7 collections.

Reviewed By: dulinriley

Differential Revision: D27695685

fbshipit-source-id: 090f6c8e8623f4a951f0b6754a93e39039172b06