Remove GC + PointerBase versions of encode functions
Summary:
When initially implementing HV32, I made `encodeNumberValue` and by
extension, `encodeHermesValue`, take both a `GC` and a `PointerBase`,
under the premise that we must already have access to a `GC*` at any
place where we need to write to a `GCHermesValue`. However, as the
implementation progressed, it turned out the right place to convert
HVs was actually further away from the calls to
`GCHermesValue::set`.

Given that there hasn't been any need for these overloads, remove
them as a simplification.

Reviewed By: dulinriley

Differential Revision: D27731080

fbshipit-source-id: dbd43f0a47f081d712e34f69052d7699e3fa8271