Force compaction in Hades when an OG collection is forced
Summary:
If `runtime->getHeap().collect(...)` is called, it is typically used by
something that wants to see memory go down as much as possible,
for example by a memory warning.

Add a `forceCompaction` argument in Hades to ensure compaction is started
in this case, since it can have a noticeable impact on fragmentation in the heap,
which would be reported by the OS as taking up memory.

Since `forceCompaction` only prepares a compactee candidate, but waits until
the next YG to evac the candidate, have `collect` also start a second YG collection.
This will only evac the compactee because YG would be empty.

It should have no impact on non-memory warning cases.

Reviewed By: neildhar

Differential Revision: D27777298

fbshipit-source-id: b08507ff6a64a015adc9405b70a978e009ba1025