Fix bug with compacting a segment with a FreelistCell and memory tracking
Summary:
The `untrackObject` API requires the object to have previously been tracked.
FreelistCells are already untracked because they don't have any JS values
stored in them, so they should be skipped.

This only applies to the compactee because YG has no free list cells.

This bug was found by forcing compaction to happen with `collect` calls, but it could've
easily happened in the wild as well after a compaction is prepared.

Reviewed By: neildhar

Differential Revision: D27780315

fbshipit-source-id: 30b3c843387b3fdb8e39af9d38c671251d1e887a