Handle untrusted doubles in JSI values
Summary:
A NaN in a JSI value does not necessarily have to be the quiet NaN,
so we need to check and convert it if necessary.

This has likely been the cause of a few crashes, since bad NaN values
could look like object pointers if they are treated as
`HermesValue`s. However, they would have been really hard to diagnose
since they would end up as bad values on the heap that then later
cause a crash somewhere else (like in the GC).

However, since HV32 needs to compress every pointer, the crash ended
up occurring at the compression step instead, which gave us a clear
stack trace pointing to external doubles as the culprit.

Reviewed By: tmikov, avp

Differential Revision: D27960846

fbshipit-source-id: bb9881d0c82f70e68edfa05c0a31b03d67cd2300