Skip registering domains for Loom
Summary:
There are two possible ways to try walking the runtime stack:
1. The background sampling thread for the engine can call
`walkRuntimeStack`. This is normally used for profiling, and the
stack (consisting of bytecode locations and `RuntimeModule`s) is
saved in the per-runtime profiler state to be symbolicated later. To
do this, we need to preserve the `Domain`s that we find in the
current stack until the trace is dumped.
2. An app wide profiler can invoke our `collectStackForLoom` method,
which will walk the stack and just retrieve the bytecode offsets and
module IDs. These have to then be symbolicated later through a
separate pipeline, since the sourcemaps are not typically available
in-app.

For the latter case, we do not need to save the domains. In fact,
saving the domains is a bug because:
1. It retains domains indefinitely, preventing any domain that has
been found in a stack trace from being freed.
2. The Loom callback cannot acquire the `runtimeDataLock_` which is
used to protect the `domains_` vector. So we cannot guarantee mutual
exclusion between the signal handler the Runtime accessing `domains_`.

Reviewed By: avp

Differential Revision: D27918611

fbshipit-source-id: 0a4a15a1355d2e09b59affe8ebda4ea761a015e1