Use ArrayStorage for GCSizingTest
Summary:
The overall motivation is to stop introducing custom cell types in
GC tests, so that we can establish a one-to-one correspondence
between CellKinds and VTables. That will allow some useful
optimisations, including allowing us to pack both the CellKind
and size into the current `VTable*` slot.

We currently maintain a custom `Array` type to be used exclusively in
unit tests. It can be swapped out almost exactly for `ArrayStorage`,
which is more straightforward to use since it is globally available
and doesn't rely on overriding metadata.

`Array` is also tagged as a `FillerCell`, which can actually cause
unexpected interactions in the GC. For instance, Hades uses
`FillerCell`s when doing cell trimming, and so won't correctly
untrack them.

This diff replaces the `Array` with `ArrayStorage` in `GCSizingTest`.

Reviewed By: dulinriley

Differential Revision: D27992233

fbshipit-source-id: d99e260ffe778ef8bc50c0576831055df637b77c