Only record segment footprint in GC stats
Summary:
Only record the memory used by segments when setting the heap size,
since the external memory is tracked separately anyway. This avoids
needing to access `ygExternalBytes_` from the background thread.

This diff does not update GenGC since it doesn't expose a
straightforward way to get the same effect. That's because GenGC uses
the `effectiveEnd` of a segment to track external memory usage, so
it's closely linked to the memory used by segments. Given that GenGC
is not used frequently, that seems fine.

Reviewed By: dulinriley

Differential Revision: D28264386

fbshipit-source-id: 2a1e7c67a900d398e4bb707f49f00230b17f98b9