Parse TS 'as' expression.
Summary:
TS can use `x as T` to typecast in situations when the type assertion
`<T>x` would be unusable due to conflicting with JSX syntax.

To support this, we add special cases to the `parseBinaryExpression`
function, which will allow us to pretend that the `as` identifier
is a binary operator in TS mode.

Reviewed By: tmikov

Differential Revision: D27239784

fbshipit-source-id: 1382ca26ae919352c3242d83544f1b5c4f7ad3b3