Make a CompressedPointer base class
Summary:
There are many places (such as `WeakRoot` and `HermesValue32`) where
we use conversion functions from `GCPointerBase` when we actually
just need to manipulate compressed pointers.

This diff factors out that common functionality into a new
`CompressedPointer` base class, that we can use everywhere we want to
perform compressed pointer operations.

The base class is meant to work similarly to `HermesValue`, where
subclasses define how it is assigned and accessed.

Reviewed By: dulinriley

Differential Revision: D28407476

fbshipit-source-id: 215babfb9ad9502c34c665b5efdcf20ffe262cc0