Use GCPointer for JSRegExp pattern
Summary:
A later diff shrinks the size of `GCCell` to 4 bytes using compressed
pointers, and reduces the number of direct property slots in
`JSObject` from 6 to 5. This causes two problems:

1. The size of `JSObject` (excluding direct slots) is now aligned on
4 bytes. This ends up creating padding in `JSRegExp` which holds a
native pointer and is therefore aligned on 8 bytes.
2. There is now one less direct property slot that we can use to
store things.

To solve (1), this diff moves the `pattern` internal property of
`JSRegExp` out of internal property slots and into a `GCPointer`
stored within the object. That lets us control the object layout more
finely than the overlap slot mechanism can (since overlap slots are
calculated using the size of the object after padding).

To solve (2), we have to allow the `lastIndex` property to go into
indirect storage.

Reviewed By: dulinriley

Differential Revision: D28407478

fbshipit-source-id: f9a2585020f9b29ba4103f57a6c2d7633e1c1eb1