Bump gradle version (#518)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/518

This diff updates the minimum gradle plugin version for building the
Hermes OSS release in order to avoid problems in older versions of
the Android NDK.

Specifically, there was a bug
(https://github.com/android/ndk/issues/1166) in the Android NDK that
caused the exception handling functions in `libc++` to be incorrectly
exported. Depending on link order, this can lead to  `libhermes.so`
not statically linking in `_Unwind_resume`, and expecting it instead
to be resolved dynamically to the version in `libc++`. However, if
we're linking against an RN version that was built against the new
version of the NDK, those symbols may not be exported, and end up
unresolved (which crashes the app).

To allow Hermes to be built with newer versions of the NDK, we need
to bump up the gradle plugin version to 4.1. RN already uses 4.1.

Bumping the gradle version seems to break the intl tests unless we
add `useLibrary 'android.test.base'` to the gradle file, so do that.
(we're technically supposed to have that anyway)

Reviewed By: mhorowitz

Differential Revision: D28763472

fbshipit-source-id: 00ffcfa94ee83a9dee0ac894e20e83867200865b