Fix for stack-overflow in isConstructor in Operations.cpp
Summary:
Return a stack overflow exception in Operations.cpp when
maximum stack depth exceeded

This functionality was added to deal with the segmentation
fault that was occurring when a deeply nested proxy is
constructed. The new added functionality throws a
RangeError, with the following error message: "Maximum
call stack size exceeded (native stack depth)".
#utd-hermes-ignore-wasm

Reviewed By: avp

Differential Revision: D28761900

fbshipit-source-id: 1c10a7e62ba497efe96efdd080172a273e0ad38f