Emit AllocObjectLiteralInst directly during IRGen
Summary:
This Diff adds a new IR node `AllocObjectLiteralInst`, which is emitted in
replace of `AllocObjectInst` when the following conditions are true:
 - object has at least one property.
 - there is no computed property keys.
 - there is no spread element nodes.
 - there is no `__proto__` assignment.
 - there is no accessors (however, if accessors are eventually overwritten to be non-accessors, it is ok). Note that a partial `HBCAllocObjectFromBufferInst` will still be created by `LowerAllocObject` pass when accessors are present.

Later on, `LowerAllocObjectLiteral` pass lowers `AllocObjectLiteralInst` to a combination of `HBCAllocObjectFromBufferInst`, `AllocObjectInst`, `StoreNewOwnPropertyInst` and `StorePropertyInst`.

`LowerNumericProperties` pass has been extended to handle `AllocObjectLiteralInst`.

A number of tests have been added/modified to work with the newly added IR node.

Documented new IR node in `doc/IR.md`.

Reviewed By: tmikov

Differential Revision: D29169308

fbshipit-source-id: 7aaa5de927b3fbe4c8ac6d4517163990b0d28d6a