Full Screen Mode enabled for the Explain Like I’m 5: Hermes video (#538)
Summary:
this will allow visitors of the site to view Explain Like I’m 5: Hermes in full-screen mode.

Currently, we cannot view Explain Like I’m 5: Hermes video in full-screen mode on the Hermes website pr will fix that.  allowfullscreen attribute was there in the code but it was not set to true. This will provide a better experience to the visitors.

Pull Request resolved: https://github.com/facebook/hermes/pull/538

Test Plan: N/A

Reviewed By: avp

Differential Revision: D29273295

Pulled By: Huxpro

fbshipit-source-id: 38e5b870ae5a4c60781e52529fdb1e87999e1e65