Avoid testing location information in time limit test.
Summary:
`exection-time-limit.js` could fail on rare occasions when compiled
with a slow version of Hermes, because the location of the exception
would be attributed to a point before the loop begins.

There's no reason for this test to check for source locations because
it's a timeout test, so just let it run and ensure the message is correct.
We can't rely on the stack trace having anything in particular in it,
because each of the functions has an AsyncBreakCheck emitted.

Reviewed By: neildhar

Differential Revision: D29311216

fbshipit-source-id: df976ee417199beec1f7bd4d06f1cd681146248a