Move StringKind from Support to BCGen/HBC
Summary: The Support directory is for code that is more or less independent of Hermes. By contrast, StringKind is intimately tied to our bytecode format, so move it into BCGen/HBC instead. One benefit of this is to make it more clear that a change to StringKind can affect the bytecode format.

Reviewed By: neildhar

Differential Revision: D29473773

fbshipit-source-id: 84c14f346acd7d6d1fc801107d524126b6019973