Use hermesc for compiling in testsuite.py
Summary:
Use `hermesc` rather than full `hermes` when compiling test cases to bytecode in testsuite.py.

Also remove some ES6 arguments that were needlessly passed on the compilation command-line.

The motivation for this change is to eventually enable testing of the compatibility between a compiler and VM of different origins.

Reviewed By: avp

Differential Revision: D29619254

fbshipit-source-id: 94a1b5a45f1b87c936865bac6202f2e0914f6120